+++ src/pages/jangmoonjung/Main/Main.js
@@ -1,40 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import Nav from '../../../components/Nav/Nav';
-import ReplyLists from './components/ReplyLists';
+import FeedLists from './components/FeedLists';
 import { FOOTER_LIST } from '../Main/components/FooterList';
 import './Main.scss';
 
 function MainJang() {
-  const [replyValue, setReplyValue] = useState('');
-  const [replyArray, setReplyArray] = useState([]);
-  const [replyDisable, setReplyDisable] = useState(true);
-  const [replyActive, setReplyActive] = useState(false);
-  const replyInput = event => {
-    setReplyValue(event.target.value);
-  };
-  const replyValidation = () => {
-    if (replyValue.trim().length > 0) {
-      setReplyActive(true);
-      setReplyDisable(false);
-    } else {
-      setReplyActive(false);
-      setReplyDisable(true);
-    }
-  };
-  const replyRegist = event => {
-    // 다른 이벤트가 발생하지 않도록
-    event.preventDefault();
-    // 빈 글일 때는 버튼 비활성화
-    if (replyValue.trim().length === 0) {
-      setReplyActive(false);
-      setReplyDisable(true);
-      return;
-    } else {
-      setReplyArray([...replyArray, replyValue]);
-      setReplyValue('');
-    }
-  };
   const [feedsInfoList, setFeedsInfoList] = useState([]);
+
   useEffect(() => {
     fetch('/data/feedsData.json').then(response => {
       response.json().then(result => setFeedsInfoList(result));
@@ -47,196 +19,8 @@ function MainJang() {
       <main className="main">
         <div className="container">
           <div className="feeds">
-            {/* <article className="container">
-              <div className="header">
-                <div className="profile">
-                  <img
-                    className="profile__img"
-                    src="/images/jangmoonjung/feeds__profile.jpg"
-                    alt="feeds__profile__img"
-                  />
-                  <div className="id">canon_mj</div>
-                </div>
-                <div className="more">
-                  <i className="more__icon fas fa-ellipsis-h" />
-                </div>
-              </div>
-              <img
-                className="feeds__img"
-                src="/images/jangmoonjung/feeds__img.jpg"
-                alt="feeds__img"
-              />
-              <div className="contents">
-                <div className="icons">
-                  <div className="left">
-                    <i className="heart far fa-heart" />
-                    <i className="comment fa-regular fa-comment" />
-                    <i className="plane fa-regular fa-paper-plane" />
-                  </div>
-                  <div className="right">
-                    <i className="bookmark fa-regular fa-bookmark" />
-                  </div>
-                </div>
-                <div className="likes">
-                  <div className="image">
-                    <img
-                      className="likes__img"
-                      src="/images/jangmoonjung/likes__profile.jpg"
-                      alt="feeds__likes__img"
-                    />
-                  </div>
-                  <div className="text">
-                    <span className="id">aineworld</span>님 외
-                    <span className="people">10명</span>이 좋아합니다
-                  </div>
-                </div>
-                <div className="texts">
-                  <div className="content">
-                    <div className="id">canon_mj</div>
-                    <div className="text" />
-                    위워크에서 진행한 베이킹 클래스...
-                  </div>
-                  <div className="more__gray">더보기</div>
-                </div>
-                <div className="newReply">
-                  <div className="texts__reply">
-                    <div className="texts__reply__left">
-                      <div className="id">neceosecius</div>
-                      <div className="text">거봐 좋았잖아~~~~😛</div>
-                    </div>
-                    <div className="texts__reply__right">
-                      <i className="texts__heart fa-regular fa-heart" />
-                      <i className="texts__remove fa-regular fa-trash-can" />
-                    </div>
-                  </div>
-                  {replyArray.map((value, index) => {
-                    return <ReplyLists key={index} content={value} />;
-                  })}
-                </div>
-                <div className="time">
-                  <span className="more__gray time__before">42분전</span>
-                </div>
-              </div>
-              <form onSubmit={replyRegist} className="reply">
-                <input
-                  value={replyValue}
-                  onChange={e => {
-                    replyInput(e);
-                    replyValidation();
-                  }}
-                  className="input"
-                  type="text"
-                  placeholder="댓글 달기...."
-                />
-                <button
-                  onClick={replyRegist}
-                  disabled={replyDisable}
-                  type="button"
-                  className={replyActive ? 'textBtnActive' : 'textBtnUnactive'}
-                >
-                  게시
-                </button>
-              </form>
-            </article> */}
             {feedsInfoList.map(feedsInfo => {
-              return (
-                <article key={feedsInfo.id} className="container">
-                  <div className="header">
-                    <div className="profile">
-                      <img
-                        className="profile__img"
-                        src={feedsInfo.profileImg_url}
-                        alt="feeds__profile__img"
-                      />
-                      <div className="id">{feedsInfo.feedId}</div>
-                    </div>
-                    <div className="more">
-                      <i className="more__icon fas fa-ellipsis-h" />
-                    </div>
-                  </div>
-                  <img
-                    className="feeds__img"
-                    src={feedsInfo.feedImg_url}
-                    alt="feeds__img"
-                  />
-                  <div className="contents">
-                    <div className="icons">
-                      <div className="left">
-                        <i className="heart far fa-heart" />
-                        <i className="comment fa-regular fa-comment" />
-                        <i className="plane fa-regular fa-paper-plane" />
-                      </div>
-                      <div className="right">
-                        <i className="bookmark fa-regular fa-bookmark" />
-                      </div>
-                    </div>
-                    <div className="likes">
-                      <div className="image">
-                        <img
-                          className="likes__img"
-                          src={feedsInfo.likeImg_url}
-                          alt="feeds__likes__img"
-                        />
-                      </div>
-                      <div className="text">
-                        <span className="id">{feedsInfo.likeId}</span>님{' '}
-                        <span className="people">외 {feedsInfo.likeNums}</span>
-                        이 좋아합니다
-                      </div>
-                    </div>
-                    <div className="texts">
-                      <div className="content">
-                        <div className="id">{feedsInfo.feedId}</div>
-                        <div className="text" />
-                        {feedsInfo.feedContent}
-                      </div>
-                      <div className="more__gray">더보기</div>
-                    </div>
-                    <div className="newReply">
-                      <div className="texts__reply">
-                        <div className="texts__reply__left">
-                          <div className="id">{feedsInfo.replyId}</div>
-                          <div className="text">{feedsInfo.replyContent}</div>
-                        </div>
-                        <div className="texts__reply__right">
-                          <i className="texts__heart fa-regular fa-heart" />
-                          <i className="texts__remove fa-regular fa-trash-can" />
-                        </div>
-                      </div>
-                      {replyArray.map((value, index) => {
-                        return <ReplyLists key={index} content={value} />;
-                      })}
-                    </div>
-                    <div className="time">
-                      <span className="more__gray time__before">
-                        {feedsInfo.timeBefore}
-                      </span>
-                    </div>
-                  </div>
-                  <form onSubmit={replyRegist} className="reply">
-                    <input
-                      value={replyValue}
-                      onChange={e => {
-                        replyInput(e);
-                        replyValidation();
-                      }}
-                      className="input"
-                      type="text"
-                      placeholder="댓글 달기...."
-                    />
-                    <button
-                      onClick={replyRegist}
-                      disabled={replyDisable}
-                      type="button"
-                      className={
-                        replyActive ? 'textBtnActive' : 'textBtnUnactive'
-                      }
-                    >
-                      게시
-                    </button>
-                  </form>
-                </article>
-              );
+              return <FeedLists key={feedsInfo.id} feedsInfo={feedsInfo} />;
             })}
           </div>
           <aside className="main-right">
